+++ src/main/java/com/graficoeweb/api/controller/MedicoController.java
@@ -1,6 +1,6 @@
 package com.graficoeweb.api.controller;
 
-import com.graficoeweb.api.medico.*;
+import com.graficoeweb.api.domain.medico.*;
 import jakarta.transaction.Transactional;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,8 +10,6 @@ import org.springframework.data.web.PageableDefault;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.util.UriComponentsBuilder;
-
-import java.util.List;
 
 @RestController
 @RequestMapping("medicos")
