+++ src/main/java/com/graficoeweb/api/controller/PacienteController.java
@@ -1,6 +1,6 @@
 package com.graficoeweb.api.controller;
 
-import com.graficoeweb.api.paciente.*;
+import com.graficoeweb.api.domain.paciente.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
