+++ src/main/java/com/graficoeweb/api/domain/paciente/DadosCadastroPaciente.java
@@ -1,6 +1,6 @@
-package com.graficoeweb.api.paciente;
+package com.graficoeweb.api.domain.paciente;
 
-import com.graficoeweb.api.endereco.DadosEndereco;
+import com.graficoeweb.api.domain.endereco.DadosEndereco;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
