+++ src/main/java/com/graficoeweb/api/domain/paciente/DadosListagemPaciente.java
@@ -1,4 +1,4 @@
-package com.graficoeweb.api.paciente;
+package com.graficoeweb.api.domain.paciente;
 public record DadosListagemPaciente(Long id, String nome, String email, String cpf) {
     public DadosListagemPaciente(Paciente paciente) {
         this(paciente.getId(), paciente.getNome(), paciente.getEmail(), paciente.getCpf());
