+++ src/main/resources/application.properties
@@ -1,7 +1,9 @@
 spring.datasource.url=jdbc:mysql://localhost/nome_banco
 spring.datasource.username=seu_user
 spring.datasource.password=sua_senha
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
+
+server.error.include-stacktrace=never
